requirements.txt
altgraph==0.17.3
auto-py-to-exe==2.26.1
bottle==0.12.23
bottle-websocket==0.2.9
Eel==0.14.0
future==0.18.2
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.1
Pillow==9.4.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
pyparsing==3.0.9
pytube==12.1.2
tk==0.1.0
whichcraft==0.6.1
zope.event==4.6
zope.interface==5.5.2
